enh: reintroduce model management to models settings
